# app/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Notification:
    properties:
      client_id:
        type: integer
      id:
        type: integer
      message:
        type: string
      notification_type:
        type: string
      sent_at:
        type: string
      status:
        type: string
    type: object
  models.Schedule:
    properties:
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      schedule_day:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления GoGRAFF
  title: GoGRAFF API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Возвращает список всех уведомлений
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Notification'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить все уведомления
      tags:
      - Уведомления
    post:
      consumes:
      - application/json
      description: Создает новое уведомление
      parameters:
      - description: Данные уведомления
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Notification'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать уведомление
      tags:
      - Уведомления
  /notifications/{id}:
    delete:
      description: Удаляет уведомление по ID
      parameters:
      - description: ID уведомления
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удалить уведомление
      tags:
      - Уведомления
    get:
      description: Возвращает уведомление по его ID
      parameters:
      - description: ID уведомления
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notification'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Уведомление не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить уведомление по ID
      tags:
      - Уведомления
    put:
      consumes:
      - application/json
      description: Обновляет данные уведомления по ID
      parameters:
      - description: ID уведомления
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные уведомления
        in: body
        name: notification
        required: true
        schema:
          $ref: '#/definitions/models.Notification'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Notification'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Уведомление не найдено
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновить уведомление
      tags:
      - Уведомления
  /schedules:
    get:
      description: Возвращает список всех расписаний
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Schedule'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить все расписания
      tags:
      - Расписания
    post:
      consumes:
      - application/json
      description: Создает новое расписание
      parameters:
      - description: Данные расписания
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Schedule'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Создать расписание
      tags:
      - Расписания
  /schedules/{id}:
    delete:
      description: Удаляет расписание по ID
      parameters:
      - description: ID расписания
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удалить расписание
      tags:
      - Расписания
    get:
      description: Возвращает данные расписания по ID
      parameters:
      - description: ID расписания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Schedule'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Расписание не найдено
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить расписание
      tags:
      - Расписания
    put:
      consumes:
      - application/json
      description: Обновляет данные расписания по ID
      parameters:
      - description: ID расписания
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные расписания
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Schedule'
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Расписание не найдено
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновить расписание
      tags:
      - Расписания
  /schedules/filter:
    get:
      description: Возвращает расписания для указанного пользователя
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Schedule'
            type: array
        "400":
          description: Некорректный запрос
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Фильтрация расписаний по пользователю
      tags:
      - Расписания
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"